refactor: Judge 도메인 생성
