--- src/main/java/utils/Judge.java
@@ -1,8 +1,0 @@
-package utils;
-
-public class Judge {
-
-    public static boolean isAble(int number) {
-        return number >= 4;
-    }
-}
